Organizes components inside individual board page and adds selections tools for each layer to recolour and delete
