+++ app/board/[boardId]/_components/board-layout/board-info.tsx
