+++ app/board/[boardId]/_components/layer-tools.tsx
@@ -1,0 +1,15 @@
+"use client";
+
+import { Camera, Colour } from "@/types/layer";
+import { memo } from "react";
+
+interface LayerToolsProps {
+  camera: Camera;
+  setLastColour: (colour: Colour) => void;
+}
+
+export const LayerTools = memo(({ camera, setLastColour }: LayerToolsProps) => {
+  return <div></div>;
+});
+
+LayerTools.displayName = "LayerTools";
